+++ lib/smtp.h
@@ -47,8 +47,8 @@ typedef enum {
 struct smtp_conn {
   struct pingpong pp;
   char *domain;    /* what to send in the EHLO */
-  int eob;         /* number of bytes of the EOB (End Of Body) that has been
-                      received thus far */
+  size_t eob;         /* number of bytes of the EOB (End Of Body) that has been
+                         received thus far */
   smtpstate state; /* always use smtp.c:state() to change state! */
   struct curl_slist *rcpt;
 };
